Update resulting forms
