+++ packages/frontend/src/components/userActious/claim/AwaitingResults.tsx
@@ -1,8 +1,13 @@
 import { FormHeading, FormText, FormWideWrapper } from '@/components/form'
+import { styled } from 'styled-components'
 
 export const AwaitingResults = () => (
-  <FormWideWrapper>
+  <BidAwaitingWrapper>
     <FormHeading>Wait for results ⏳</FormHeading>
     <FormText>The bidding window has closed. Waiting for the organizers to settle the contest.</FormText>
-  </FormWideWrapper>
+  </BidAwaitingWrapper>
 )
+
+const BidAwaitingWrapper = styled(FormWideWrapper)`
+  width: 460px;
+`
